pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "novarl"
version = "0.0.1"
description = "A modular reinforcement learning training framework for language model alignment."
authors = [{name = "NovaRL"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pre-commit>=3.5",
    "black>=23.9",
    "ruff>=0.1.6",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
